Defined constants as static final and fixed build issue

+++ core/src/com/mygdx/gigabiteconomy/sprites/IHealthBar.java
@@ -5,6 +5,6 @@ import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 public interface IHealthBar {
     void drawOn(SpriteBatch batch);
 
-    void modifyHealth(int dhealth);
+    void modifyHealth(float dhealth);
 
 }
